Simplify packaging (fix LP#1784418)
The problem was caused by two files: VERSION and compile_translations.py.
Neither should be in a bdist, but both are required in the sdist.

It’s tricky to fix this bug directly, so instead remove the need for these
files: fold them both into setup.py.

Also, remove share/icons/hicolor/index.theme, which served no purpose, and
as a result replace some code in setup.py with a single-line MANIFEST.in.